# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library Management API",
    "description": "A simple REST API for managing books in a library.",
    "version": "1.0"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "responses": {
          "200": {
            "description": "Returns a list of all books",
            "schema": {
              "type": "object",
              "properties": {
                "books": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "The unique ID of the book"
                      },
                      "book": {
                        "type": "object",
                        "description": "Details of the book",
                        "properties": {
                          "title": {
                            "type": "string",
                            "description": "Title of the book"
                          },
                          "author": {
                            "type": "string",
                            "description": "Author of the book"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new book",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Book object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "description": "Title of the book"
                },
                "author": {
                  "type": "string",
                  "description": "Author of the book"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Book added successfully",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "The unique ID of the added book"
                },
                "book": {
                  "type": "object",
                  "description": "Details of the added book",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "Title of the added book"
                    },
                    "author": {
                      "type": "string",
                      "description": "Author of the added book"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}": {
      "get": {
        "summary": "Get details of a specific book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "description": "ID of the book to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the requested book",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "The unique ID of the book"
                },
                "book": {
                  "type": "object",
                  "description": "Details of the book",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "Title of the book"
                    },
                    "author": {
                      "type": "string",
                      "description": "Author of the book"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "summary": "Update details of a specific book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "description": "ID of the book to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated book object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "description": "Updated title of the book"
                },
                "author": {
                  "type": "string",
                  "description": "Updated author of the book"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "The unique ID of the updated book"
                },
                "book": {
                  "type": "object",
                  "description": "Details of the updated book",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "Updated title of the book"
                    },
                    "author": {
                      "type": "string",
                      "description": "Updated author of the book"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "description": "ID of the book to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    },
    "/config": {
      "get": {
        "summary": "Get environment variables",
        "description": "Returns a JSON object containing environment variables.",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/fib": {
      "get": {
        "summary": "Get Fibonacci sequence",
        "description": "Returns a JSON array containing Fibonacci sequence.",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "Length of the Fibonacci sequence",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      }
    }    
  }
}
 
  